Meet a Candidate now gets a random list of candidates
Added a new API route at /api/candidates/random_list that returns a
randomly-sorted list of candidate RCS IDs. Reworked HomeController to use this
endpoint to iterate through random candidates.

This resolves #73. Additionally, all offices that a candidate is running for
are now displayed simultaneuously in the Meet a Candidate widget.