Some Updates based on GM Week 2019 (#107)
* Add node SESSION_SECRET to config.js

* Display Graduate Nominator Status

The entry/exit date does not matter for grads

* Remove candidate admin rights

This overrides the CMS API call for candidates who happen to be members 
of the EC and WTG on CMS. Once somebody is added to the website they 
give up their admin rights until they are removed as a candidate. One 
thing to note is that I use a synchronous MySQL library instead of the 
default async one.